Dynamically creating one sifter menu item on watch. Need to modify to create all menu items.

+++ Pebble/pebble_sifter/src/node.h
@@ -11,4 +11,4 @@ void freeNodes(struct node *root);
 
 struct node* getNodeAtIndex(struct node *root, int index);
 
-size_t size(struct node *root);
+int size(struct node *root);
